feat: reorganize document types and messages into a dedicated types file
